# Improve performance of hashing `SmallBitVec`s